ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0

opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78

deepface==0.0.79

numpy>=1.26.0

Pillow==10.1.0

SpeechRecognition==3.10.0
pydub==0.25.1

# transformers==4.35.2
# torch==2.1.0
# sentence-transformers==2.2.2

# spacy==3.7.2
scikit-learn==1.3.2

# tensorflow==2.15.0

gunicorn==21.2.0
python-dotenv==1.0.0